initial combine events macro
