+++ singularity_common/src/lib.rs
@@ -5,5 +5,3 @@ pub mod project;
 pub mod sap;
 pub mod tab;
 pub mod utils;
-
-pub mod macro_reexports;
