--- singularity_common/src/macro_reexports.rs
@@ -1,6 +1,0 @@
-pub use singularity_macros::combine_events_macro;
-
-#[macro_export]
-macro_rules! combine_events {
-    {$inner:tt} => {$crate::macro_reexports::combine_events_macro!(inner);}
-}
