+++ singularity_common/src/sap/mod.rs
@@ -1,3 +1,6 @@
 pub mod byte_stream;
+pub mod packet;
+
+// TODO: the two dependencies below should be removed
 pub mod standard_features;
 pub mod universal_packet;
